// schemas/setting.ts
export const financeSettingSchema = z
  .object({
    base_CurrencyId: z.number().min(0, "Base Currency is required"),
    local_CurrencyId: z.number().min(0, "Local Currency is required"),
    exhGain_GlId: z.number().min(0, "Exchange Gain Account is required"),
    exhLoss_GlId: z.number().min(0, "Exchange Loss Account is required"),
    bankCharge_GlId: z.number().min(0, "Bank Charges Account is required"),
    adjCharge_GlId: z.number().min(0).optional(),
    profitLoss_GlId: z.number().min(0, "Profit & Loss Account is required"),
    retEarning_GlId: z.number().min(0, "Retained Earnings Account is required"),
    saleGst_GlId: z.number().min(0, "Sales GST Account is required"),
    purGst_GlId: z.number().min(0, "Purchase GST Account is required"),
    saleDef_GlId: z.number().min(0, "Sales Deferred Account is required"),
    purDef_GlId: z.number().min(0, "Purchase Deferred Account is required"),
  })
  .partial()

export type FinanceSettingSchemaType = z.infer<typeof financeSettingSchema>